refactor:move resource view components to under the ResourceViews/List/ directory
