+++ web/components/ResourceViews/List/UnscheduledPodList.vue
@@ -12,8 +12,8 @@ import {
   onMounted,
   defineComponent,
 } from "@nuxtjs/composition-api";
-import {} from "./lib/util";
-import PodStoreKey from "./StoreKey/PodStoreKey";
+import {} from "../../lib/util";
+import PodStoreKey from "../../StoreKey/PodStoreKey";
 export default defineComponent({
   setup() {
     const store = inject(PodStoreKey);
