+++ web/pages/index.vue
@@ -38,12 +38,12 @@ import PVCStoreProvider from "~/components/StoreProvider/PVCStoreProvider.vue";
 import SchedulerConfigurationStoreProvider from "~/components/StoreProvider/SchedulerConfigurationStoreProvider.vue";
 import StorageClassStoreProvider from "~/components/StoreProvider/StorageClassStoreProvider.vue";
 import PriorityClassStoreProvider from "~/components/StoreProvider/PriorityClassStoreProvider.vue";
-import NodeList from "~/components/NodeList.vue";
-import UnscheduledPodList from "~/components/UnscheduledPodList.vue";
-import PVList from "~/components/PVList.vue";
-import PVCList from "~/components/PVCList.vue";
-import StorageClassList from "~/components/StorageClassList.vue";
-import PriorityClassList from "~/components/PriorityClassList.vue";
+import NodeList from "~/components/ResourceViews/List/NodeList.vue";
+import UnscheduledPodList from "~/components/ResourceViews/List/UnscheduledPodList.vue";
+import PVList from "~/components/ResourceViews/List/PVList.vue";
+import PVCList from "~/components/ResourceViews/List/PVCList.vue";
+import StorageClassList from "~/components/ResourceViews/List/StorageClassList.vue";
+import PriorityClassList from "~/components/ResourceViews/List/PriorityClassList.vue";
 import PodStoreProvider from "~/components/StoreProvider/PodStoreProvider.vue";
 import SnackbarStoreProvider from "~/components/StoreProvider/SnackbarStoreProvider.vue";
 import ResourceAddButton from "~/components/ResourceAddButton.vue";
